requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyfgc-const==0.0.1
pyfgc-decoders==0.0.3
pyfgc-name==1.3.1
pyfgc-rbac==1.1.0
pyparsing==2.4.7
pyserial==3.4
pytest==5.4.2
pytest-asyncio==0.12.0
requests==2.22.0
six==1.14.0
urllib3==1.25.7
wcwidth==0.1.9
zipp==3.1.0
